feat(rdap): 🚦 use enums for RDAP statuses
