+++ DomainDetective.Tests/TestRdapModels.cs
@@ -7,12 +7,13 @@ public class TestRdapModels
     [Fact]
     public void DomainSerializationRoundTrip()
     {
-        const string json = "{\"ldhName\":\"example.com\",\"nameservers\":[{\"ldhName\":\"ns1.example.net\"},{\"ldhName\":\"ns2.example.net\"}],\"events\":[{\"eventAction\":\"registration\",\"eventDate\":\"2000-01-01T00:00:00Z\"},{\"eventAction\":\"expiration\",\"eventDate\":\"2030-01-01T00:00:00Z\"}],\"entities\":[{\"handle\":\"123\",\"roles\":[\"registrar\"],\"vcardArray\":[\"vcard\",[[\"fn\",{},\"text\",\"Registrar Inc\"]]]}]}";
+        const string json = "{\"ldhName\":\"example.com\",\"status\":[\"active\"],\"nameservers\":[{\"ldhName\":\"ns1.example.net\"},{\"ldhName\":\"ns2.example.net\"}],\"events\":[{\"eventAction\":\"registration\",\"eventDate\":\"2000-01-01T00:00:00Z\"},{\"eventAction\":\"expiration\",\"eventDate\":\"2030-01-01T00:00:00Z\"}],\"entities\":[{\"handle\":\"123\",\"roles\":[\"registrar\"],\"vcardArray\":[\"vcard\",[[\"fn\",{},\"text\",\"Registrar Inc\"]]]}]}";
 
         var domain = JsonSerializer.Deserialize<RdapDomain>(json, RdapJson.Options)!;
         Assert.Equal("example.com", domain.LdhName);
         Assert.Equal("ns1.example.net", domain.Nameservers[0].LdhName);
         Assert.Equal(RdapEventAction.Registration, domain.Events[0].Action);
+        Assert.Equal(RdapDomainStatus.Active, domain.Status[0]);
 
         var serialized = JsonSerializer.Serialize(domain, RdapJson.Options);
         var round = JsonSerializer.Deserialize<RdapDomain>(serialized, RdapJson.Options)!;
@@ -44,11 +45,12 @@ public class TestRdapModels
     [Fact]
     public void EntityParsing()
     {
-        const string json = "{\"handle\":\"ABC123\",\"roles\":[\"registrant\"],\"vcardArray\":[\"vcard\",[[\"fn\",{},\"text\",\"John Doe\"]]]}";
+        const string json = "{\"handle\":\"ABC123\",\"roles\":[\"registrant\"],\"status\":[\"validated\"],\"vcardArray\":[\"vcard\",[[\"fn\",{},\"text\",\"John Doe\"]]]}";
         var entity = JsonSerializer.Deserialize<RdapEntity>(json, RdapJson.Options)!;
         Assert.Equal("ABC123", entity.Handle);
         Assert.Contains("registrant", entity.Roles);
         Assert.True(entity.VcardArray.HasValue);
+        Assert.Contains(RdapDomainStatus.Validated, entity.Status);
     }
 
     [Fact]
